// bconnect-compliance-mcp/src/generated/compliance-types.ts
export interface components {
    schemas: {
// ...
        DetectedVulnerabilityPagedList: {
            /** Format: int32 */
            readonly currentPage?: number | null;
            /** Format: int32 */
            readonly pageSize?: number | null;
            /** Format: int32 */
            readonly totalPages?: number | null;
            /** Format: int32 */
            readonly totalItems?: number;
            readonly hasPreviousPage?: boolean;
            readonly hasNextPage?: boolean;
            readonly data?: components["schemas"]["DetectedVulnerability"][];
        };
// ...
        Rule: {
            /**
             * Format: guid
             * @description The identifier of the rule
             */
            id?: string;
            /** @description The name of the rule */
            ruleName: string;
            /** @description The rule type */
            type: string;
            /** @description The Severity of the rule */
            severity: string;
            /** @description Description of the rule */
            description?: string;
        };
        RulePagedList: {
            /** Format: int32 */
            readonly currentPage?: number | null;
            /** Format: int32 */
            readonly pageSize?: number | null;
            /** Format: int32 */
            readonly totalPages?: number | null;
            /** Format: int32 */
            readonly totalItems?: number;
            readonly hasPreviousPage?: boolean;
            readonly hasNextPage?: boolean;
            readonly data?: components["schemas"]["Rule"][];
        };
// ...
        RuleViolationPagedList: {
            /** Format: int32 */
            readonly currentPage?: number | null;
            /** Format: int32 */
            readonly pageSize?: number | null;
            /** Format: int32 */
            readonly totalPages?: number | null;
            /** Format: int32 */
            readonly totalItems?: number;
            readonly hasPreviousPage?: boolean;
            readonly hasNextPage?: boolean;
            readonly data?: components["schemas"]["RuleViolation"][];
        };
        Vulnerability: {
            /**
             * Format: guid
             * @description The identifier of cve
             */
            id?: string;
            /** @description The cve id */
            cveId: string;
            /**
             * Format: double
             * @description Specifies the cvss score of cve
             */
            cvssScore: number;
            /** @description Specifies the cve severity */
            severity: string;
            /** @description Description of cve */
            description?: string | null;
            /** @description String with all affected products */
            affectedProducts?: string | null;
            /** @description String with all affected operating systems */
            affectedOperatingSystems?: string | null;
        };
        VulnerabilityPagedList: {
            /** Format: int32 */
            readonly currentPage?: number | null;
            /** Format: int32 */
            readonly pageSize?: number | null;
            /** Format: int32 */
            readonly totalPages?: number | null;
            /** Format: int32 */
            readonly totalItems?: number;
            readonly hasPreviousPage?: boolean;
            readonly hasNextPage?: boolean;
            readonly data?: components["schemas"]["Vulnerability"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    GetDetectedRuleViolations: {
        parameters: {
            query?: {
                /** @description Sorts results by property name and with sort direction. Multiple properties are separated by comma.<br />Possible values are EndpointName or RuleName (e.g. RuleName asc). */
                OrderBy?: string;
                /** @description Filters results by matching the given value against searchable properties.<br />Searchable values are EndpointName or RuleName. */
                SearchQuery?: string;
                /** @description The zero-indexed number of the first page that begins the set of pages that are returned in the response. */
                Page?: number;
                /** @description The number of items to list on a single page. For example, if page_size is 10, each page shows ten items. <br />A valid value is a non-negative, non-zero integer. The default value is 20. Maximum value is 1000. */
                PageSize?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns all detected rule violations for Android, iOS and macOS endpoints */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["RuleViolationPagedList"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description Internal Server Error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
    GetDetectedRuleViolationsForEndpoint: {
        parameters: {
            query?: {
                /** @description Sorts results by property name and with sort direction. Multiple properties are separated by comma.<br />Possible values are EndpointName or RuleName (e.g. RuleName asc). */
                OrderBy?: string;
                /** @description Filters results by matching the given value against searchable properties.<br />Searchable values are EndpointName or RuleName. */
                SearchQuery?: string;
                /** @description The zero-indexed number of the first page that begins the set of pages that are returned in the response. */
                Page?: number;
                /** @description The number of items to list on a single page. For example, if page_size is 10, each page shows ten items. <br />A valid value is a non-negative, non-zero integer. The default value is 20. Maximum value is 1000. */
                PageSize?: number;
            };
            header?: never;
            path: {
                endpointId: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns detected rule violations for an Android, iOS or macOS endpoint with specified endpoint id */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["RuleViolationPagedList"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description An Android, iOS or macOS endpoint with the specified id does not exist or contains no rule violation */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description Internal Server Error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
    GetAllDetectedVulnerabilities: {
        parameters: {
            query?: {
                /** @description Sorts results by property name and with sort direction. Multiple properties are separated by comma.<br />Possible values are EndpointName or CveId (e.g. EndpointName asc). */
                OrderBy?: string;
                /** @description Filters results by matching the given value against searchable properties.<br />Searchable values are EndpointName or CveId. */
                SearchQuery?: string;
                /** @description The zero-indexed number of the first page that begins the set of pages that are returned in the response. */
                Page?: number;
                /** @description The number of items to list on a single page. For example, if page_size is 10, each page shows ten items. <br />A valid value is a non-negative, non-zero integer. The default value is 20. Maximum value is 1000. */
                PageSize?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns all detected vulnerabilities for Windows endpoints */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["DetectedVulnerabilityPagedList"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description Internal Server Error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
// ...
}